bugfix: fix pybind class bindings (#255)
Previously we bind a factory method as the init function for C++ Class
in Pybind, which is returned by value instead of reference. The
destructive function of handlers in #253 will be triggered twice, and it
leads to segmentation faults.

This PR bypass the factory method and initializes the C++ classes
directly.