feat: expose pytorch api for block sparse attention (#375)
The block sparse attention (for any block size (R, C)) are hidden in
flashinfer's codebase but it was never exposed explicitly in python. As
requested in #367 , this PR implements the PyTorch APIs for block sparse
attention, accordingly to our experiments, it can greatly accelerate
attention computation with low density (10x for Tree Attention in
Sequoia).