bugfix: Fix the batch size & kv indptr calculation (#641)
The new_batch_size was not computed correctly for CUDA graph mode. Now
we ensure it is in sync with the contents of the per-block indices. Also
adjusted the computation of the KV chunk sizes for CUDA graph mode.

This PR also removes the `max_seq_len` parameter from kernels, as a
reasonable upper bound from it can be derived from `total_num_rows`.