sampling: simplify min-p sampling (#713)
As suggested by @aikitoria and @rolandtannous in #710 , we don't need
rejection algorithm for min-p sampling, this PR simplifies the design.

There is a breaking change on API: we no longer returns `success` array
for min-p sampling because there is no risk of rejecting the samples. In
later PRs, we will also remove the `success` array in top-p/top-k
sampling APIs.