bugfix: Fix no return type error (#904)
This pull request includes a change to the `convert_s_to_p` function in
the `mla_hopper.cuh` file to improve type safety and clarity.

Function signature modification:

*
[`include/flashinfer/attention/mla_hopper.cuh`](diffhunk://#diff-5ebeba4848db48074fb2636d9ef52f85409674d9fcaba1e2b8091cf6fc36e651L554-R554):
Changed the return type of the `convert_s_to_p` function from `auto` to
`void` to explicitly indicate that the function does not return a value.