bugfix: Fix compilation with FP16_QK_REDUCTION enabled. (#962)
As described in #806 and #936, setting the cmake build flag
`FLASHINFER_GEN_USE_FP16_QK_REDUCTIONS` to "true" causes a build failure
due to `cuda_fp16.h` not supporting `constexpr` cast from `__half` type
to `float`. Note that the issue is not just a CMake/C++ configuration
issue the issue will be triggered even in the flashinfer JIT code
compilation path as reported in #915.

The PR fixes #806 and #936 by adding a modified version of the FP16
header from the [FP16 library](https://github.com/Maratyszcza/FP16) that
supports `constexpr` versions of the conversion functions. To make the
conversion functions `constexpr`, I am using `std::bit_cast` that is the
reason for bumping the required standard to 20.

With these changes I am able to build the C++ API with both
`FLASHINFER_GEN_USE_FP16_QK_REDUCTIONS` ON and OFF.

Fixes #936 
Fixes #806