perf: add `-DNDEBUG` compilation flag (#998)
We observed some performance degradation after upgrading cutlass to
v3.9.

This is because some runtime checks are added to cutlass in debug mode:
*
https://github.com/NVIDIA/cutlass/blob/df8a550d3917b0e97f416b2ed8c2d786f7f686a3/include/cutlass/pipeline/sm90_pipeline.hpp#L66-L70
*
https://github.com/NVIDIA/cutlass/blob/df8a550d3917b0e97f416b2ed8c2d786f7f686a3/include/cutlass/pipeline/sm90_pipeline.hpp#L82-L86

This PR addresses the issue by adding `-DNDEBUG` to compilation flags in
both JIT and AOT mode.

After this PR, we no longer observe the performance degradation.