chore: remove cpp benchmarks, tests, cmake path, as they are deprecated (#1345)
After discovering that these tests and benchmarks do not build, I
attempted to repair them
[here](https://github.com/flashinfer-ai/flashinfer/pull/1336). I was
informed that this part of the code is deprecated, and that CMake is not
used to perform builds anymore.

To avoid anyone else attempting to use this code and spending time
trying to repair it, I suggest that it be removed.