fix: new issue workflow (#1773)
<!-- .github/pull_request_template.md -->

## 📌 Description

The intent of the workflow is to add `needs-triage` when a new issue is
created. The initial version of this workflow, added in #1765 failed
when I created dummy issue #1772. This PR is a fix for that bug.

## 🔍 Related Issues

<!-- Link any related issues here -->

## 🚀 Pull Request Checklist

Thank you for contributing to FlashInfer! Before we review your pull
request, please make sure the following items are complete.

### ✅ Pre-commit Checks

- [x] I have installed `pre-commit` by running `pip install pre-commit`
(or used your preferred method).
- [x] I have installed the hooks with `pre-commit install`.
- [x] I have run the hooks manually with `pre-commit run --all-files`
and fixed any reported issues.

> If you are unsure about how to set up `pre-commit`, see [the
pre-commit documentation](https://pre-commit.com/).

## 🧪 Tests

I have tested with `gh` CLI targeting my fork of this repo and also by
adding the action to my fork.

Completed action:
https://github.com/sricketts/flashinfer/actions/runs/18041739971

- [x] Tests have been added or updated as needed.
- [ ] All tests are passing (`unittest`, etc.).

## Reviewer Notes

<!-- Optional: anything you'd like reviewers to focus on, concerns, etc.
-->