bugfix: remove the append "a" logic if user specifies cuda arch explicitly (#1798)
<!-- .github/pull_request_template.md -->

## 📌 Description

In our design of compilation context we will append "a" to cuda archs if
the major arch is >= 9, however, we shouldn't change it if user specify
them explicitly.

This PR changes the logic so that we only append "a" to cuda arch if
`FLASHINFER_CUDA_ARCH_LIST` is not set explicitly, otherwise follow the
user specified cuda arch.

## 🔍 Related Issues

<!-- Link any related issues here -->

## 🚀 Pull Request Checklist

Thank you for contributing to FlashInfer! Before we review your pull
request, please make sure the following items are complete.

### ✅ Pre-commit Checks

- [x] I have installed `pre-commit` by running `pip install pre-commit`
(or used your preferred method).
- [x] I have installed the hooks with `pre-commit install`.
- [x] I have run the hooks manually with `pre-commit run --all-files`
and fixed any reported issues.

> If you are unsure about how to set up `pre-commit`, see [the
pre-commit documentation](https://pre-commit.com/).

## 🧪 Tests

- [x] Tests have been added or updated as needed.
- [ ] All tests are passing (`unittest`, etc.).

## Reviewer Notes

<!-- Optional: anything you'd like reviewers to focus on, concerns, etc.
-->